+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 1.7.0
+- Amend file path so the list shows full path.
 - Amend default import identifiers.
 - Add namespace imports to the list.
 - Add support of multiple-source filters in the extension configuration.